player and enemy;
scripts
player health
player stats
enemy melee

+++ Dinheirinho/Assets/Scripts/Enemy Scripts/EnemyHealth.cs
@@ -4,9 +4,9 @@ using UnityEngine;
 
 public class EnemyHealth : MonoBehaviour
 {
-    public int maxHealth;
+    public float maxHealth;
 
-    int currentHealth;
+    float currentHealth;
 
     void Start()
     {
